Export endpoint'ini script haline getiricsr
